Add command that resets the vote creation cooldown
When a player votes, the current time is saved. Whenever a new vote is cast, sv_vote_creation_timer is read and used to determine if enough time has passed, which means that changing sv_vote_creation_timer to vote to reset a run on map 1 will leave you without a vote at the start of the next run. The new command added by this commit solves that issue.